requirements.txt
flask
flask-cors
cloud-sql-python-connector[pymysql]
sqlalchemy
pymysql
gunicorn
pandas
openpyxl
google-generativeai>=0.5.0
